telemetry/src/test/java/com/kroger/telemetry/facet/FacetTest.kt
/**
 * MIT License
 *
 * Copyright (c) 2021 The Kroger Co. All rights reserved.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package com.kroger.telemetry.facet

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Assertions.assertTrue
import org.junit.jupiter.api.Test

internal class FacetTest {
    @Test
    fun `GIVEN computed facet THEN computations will be run multiple times`() {
        var count = 0
        val facet = object : Facet.Computed<Unit> {
            override val compute: () -> Unit = { count += 1 }
        }

        for (i in 0..2) facet.compute()
        assertTrue(count > 1)
    }

    @Test
    fun `GIVEN lazy facet THEN computations will be be run once`() {
        var count = 0
        val facet = object : Facet.Lazy<Int>() {
            override val compute: () -> Int = {
                count += 1
                count
            }
        }

        // read value twice to ensure compute is only run once
        assertEquals(1, facet.value)
        assertEquals(1, facet.value)
    }
}
